Move Pipeline Triggers to a separate file.
Creates the new service PipelineTriggersService for handling requests to the Pipeline Triggers part of the GitLab API.
